Migrating to class objects format
Migrated most of the parameters specification to the class RocketMotor. Included a geometry.py with the formulas for area and volume used. Found and fixed a mistake in the equation for the corrected exit velocity (wrong exponent, now all the corrected velocities, pressures and forces converge to the same value). Added a function to save the thrust to a CSV file. Minor alterations in other places.